refactor(email): move email-related utility functions to appointment-helpers module
